F dev migrate to dotnet standart

+++ REMME.Auth/REMME.Auth.Client/Contracts/IRemmeToken.cs
@@ -21,6 +21,6 @@ namespace REMME.Auth.Client.Contracts
         /// </summary>
         /// <param name="publicKey"></param>
         /// <returns>Integer representing REM balance</returns>
-        Task<int> GetBalance(string publicKey);
+        Task<ulong> GetBalance(string publicKey);
     }
 }
